+++ inc/main.php
@@ -59,11 +59,6 @@ function rocket_init() {
 	if ( rocket_valid_key() ) {
 		require WP_ROCKET_COMMON_PATH . 'purge.php';
 
-		if ( 0 < (int) get_rocket_option( 'do_cloudflare' ) ) {
-			require WP_ROCKET_FUNCTIONS_PATH . 'cloudflare.php';
-			require WP_ROCKET_COMMON_PATH . 'cloudflare.php';
-		}
-
 		if ( is_multisite() && defined( 'SUNRISE' ) && SUNRISE === 'on' && function_exists( 'domain_mapping_siteurl' ) ) {
 			require WP_ROCKET_INC_PATH . '/domain-mapping.php';
 		}
